Lockers/requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cmake==3.30.2
colorama==0.4.6
Django==5.1
django-bootstrap5==24.2
django-widget-tweaks==1.5.0
facenet-pytorch==2.6.0
filelock==3.15.4
fsspec==2024.6.1
idna==3.8
Jinja2==3.1.4
kombu==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.3
numpy==1.26.4
pillow==10.2.0
prompt_toolkit==3.0.47
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
sqlparse==0.5.1
sympy==1.13.2
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
